backend/requirements.txt
# ai-report-generator/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.14.0
reportlab==4.0.9
matplotlib==3.8.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx>=0.27.0
python-dateutil==2.8.2